requirements.txt
Flask>=3.0,<4.0
Pillow>=10.0,<12.0
ultralytics>=8.3,<9.0

